feat(logging): rename pacakge kitexlogrus -> logrus; warn level in default case; git ignore .idea
