+++ .gitignore
@@ -4,7 +4,7 @@
 *.dll
 *.so
 *.dylib
-
+.idea
 # Test binary, built with `go test -c`
 *.test
