+++ logging/logrus/option.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package kitexlogrus
+package logrus
 
 import (
 	"github.com/sirupsen/logrus"
